chore: reorganize getFilePaths.ts and update import paths
