+++ src/cli/update-strict-comments/updateStrictComments.ts
@@ -1,4 +1,4 @@
-import { getFilePathsWithErrors, getFilePathsOnPathWithoutErrors } from './getFilePaths';
+import { getFilePathsWithErrors, getFilePathsOnPathWithoutErrors } from '../getFilePaths';
 import { isIgnoreCommentPresent, isStrictCommentPresent } from '../isCommentPresent';
 import { isFileStrictByPath } from '../../common/isFileStrictByPath';
 import { insertIgnoreComment, removeStrictComment } from '../commentOperations';
